feat: web engine hook
