+++ apps/web/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useWebEngine } from './useWebEngine';
