+++ apps/web/pages/index.tsx
@@ -1,141 +1,27 @@
 import {
   WidgetActivityMonitor,
   WidgetMonitor,
-  WidgetUpdate,
-  WidgetDOMElement,
-  onCallbackInvocation,
-  onCallbackResponse,
-  onRender,
 } from '@bos-web-engine/application';
-import type { ComponentCompilerResponse } from '@bos-web-engine/compiler';
 import { getAppDomId, getIframeId, SandboxedIframe } from '@bos-web-engine/iframe';
-import React, { useEffect, useState } from 'react';
-import ReactDOM from 'react-dom/client';
+import React, { useState } from 'react';
+
+import { useWebEngine } from '../hooks';
 
 const DEFAULT_ROOT_WIDGET = 'andyh.near/widget/MainPage';
 
-const roots = {} as { [key: string]: ReactDOM.Root };
-const widgets = {} as { [key: string]: any };
-
 const monitor = new WidgetActivityMonitor();
-
-function mountElement({ widgetId, element }: { widgetId: string, element: WidgetDOMElement }) {
-  if (!roots[widgetId]) {
-    const domElement = document.getElementById(getAppDomId(widgetId));
-    if (!domElement) {
-      const metricKey = widgetId.split('##')[0];
-      monitor.missingWidgetReferenced(metricKey);
-      console.error(`Node not found: #${getAppDomId(widgetId)}`);
-      return;
-    }
-
-    roots[widgetId] = ReactDOM.createRoot(domElement);
-  }
-
-  roots[widgetId].render(element);
-}
 
 export default function Web() {
   const [rootWidget, setRootWidget] = useState('');
   const [rootWidgetInput, setRootWidgetInput] = useState(DEFAULT_ROOT_WIDGET);
-  const [rootWidgetSource, setRootWidgetSource] = useState<string | null>(null);
-  const [widgetUpdates, setWidgetUpdates] = useState('');
   const [showMonitor, setShowMonitor] = useState(true);
   const [showWidgetDebug, setShowWidgetDebug] = useState(true);
-  const [compiler, setCompiler] = useState<any>(null);
 
-  const widgetProxy = new Proxy(widgets, {
-    get(target, key: string) {
-      return target[key];
-    },
-
-    set(target, key: string, value: any) {
-      // if the widget is being added, initiate request for widget component code
-      if (!target[key]) {
-        compiler?.postMessage({ componentId: key, isTrusted: value.isTrusted });
-      }
-
-      target[key] = value;
-      return true;
-    },
+  const { components } = useWebEngine({
+    monitor,
+    showWidgetDebug,
+    rootWidget,
   });
-
-  useEffect(() => {
-    function buildMessageListener(eventType: string) {
-      return function (event: any) {
-        try {
-          if (typeof event.data !== 'object' || event.data.type !== eventType) {
-            return;
-          }
-
-          const { data } = event;
-          switch (eventType) {
-              case 'widget.callbackInvocation': {
-                monitor.widgetCallbackInvoked(data);
-                onCallbackInvocation({ data });
-                break;
-              }
-              case 'widget.callbackResponse': {
-                monitor.widgetCallbackReturned(data);
-                onCallbackResponse({ data });
-                break;
-              }
-              case 'widget.render': {
-                monitor.widgetRendered(data);
-                onRender({
-                  data,
-                  isDebug: showWidgetDebug,
-                  markWidgetUpdated: (update: WidgetUpdate) => monitor.widgetUpdated(update),
-                  mountElement,
-                  widgets: widgetProxy,
-                });
-                break;
-              }
-              default:
-                break;
-          }
-        } catch (e) {
-          console.error({ event }, e);
-        }
-      };
-    }
-
-    const messageListeners = [
-      buildMessageListener('widget.callbackInvocation'),
-      buildMessageListener('widget.callbackResponse'),
-      buildMessageListener('widget.render'),
-    ];
-
-    messageListeners.forEach((cb) => window.addEventListener('message', cb));
-    return () => messageListeners.forEach((cb) => window.removeEventListener('message', cb));
-  }, [rootWidgetSource, showWidgetDebug]);
-
-  useEffect(() => {
-    if (!rootWidget) {
-      return;
-    }
-
-    if (!compiler) {
-      const worker = new Worker(new URL('../workers/compiler.ts', import.meta.url));
-      setCompiler(worker);
-    } else {
-      compiler.onmessage = ({ data }: MessageEvent<ComponentCompilerResponse>) => {
-        const { componentId, componentSource } = data;
-        const component = { ...widgetProxy[componentId], widgetComponent: componentSource };
-        if (!rootWidgetSource && componentId === rootWidget) {
-          setRootWidgetSource(componentId);
-        }
-        monitor.widgetAdded({ source: componentId, ...component });
-        setWidgetUpdates(widgetUpdates + componentId);
-        widgetProxy[componentId] = component;
-      };
-
-      compiler.postMessage({
-        componentId: rootWidget,
-        isTrusted: false,
-      });
-    }
-  }, [rootWidget, rootWidgetSource, compiler]);
 
   return (
     <div className='App'>
@@ -179,8 +65,8 @@ export default function Web() {
           <div className="iframes">
             {showWidgetDebug && (<h5>here be hidden iframes</h5>)}
             {
-              Object.entries({ ...widgetProxy })
-                .filter(([, { widgetComponent }]) => !!widgetComponent)
+              Object.entries(components)
+                .filter(([, component]) => !!component?.widgetComponent)
                 .map(([widgetId, { isTrusted, props, widgetComponent }]) => (
                   <div key={widgetId} widget-id={widgetId}>
                     <SandboxedIframe
